Added the first version of SmartSensing module
This new class is the core of the new functionalities of wake up the device
in a low power mode and perform readings of a sensor.
With this release the functionalities related to the sampling befor the boot has been intergrated,
while the daemon part, which performs readings when the kernel is on, not yet.
Indeed a function debugInt has been added for the debugging purposes of see an
integer on the console before the boot of the kernel.

Signed-off-by: Alessandro Rizzi <alexander89@katamail.com>
Signed-off-by: Terraneo Federico <fede.tft@hotmail.it>