Added an interface to easily access to SmartSensing class
The function getSmartSensingDriver has been introducted in order to retrive the instance of the class easily.
Then has been defined an instace of the SmartSensing class with a fixed dimension.
This has been performed in order to simplify the code maintenance and avoid possible errors by only keeping
one instance of the class and define only once the correct class size.

Signed-off-by: Alessandro Rizzi <alexander89@katamail.com>
Signed-off-by: Terraneo Federico <fede.tft@hotmail.it>