Fixed a couple of bugs in the wakeupDaemon method of SuspendManager class
The first one regards the use of a "more-than" instead of a "less-than".
The second instead is related to the time which pass between the use of a getTick and the sleep usage.
Both of them cause the call of a sleep with a negative integer which practically never respawn the given process.

Signed-off-by: Alessandro Rizzi <alexander89@katamail.com>
Signed-off-by: Terraneo Federico <fede.tft@hotmail.it>