IRQportableStartKernel splitted
It is ensured that the miosix::kernel_started flag is false during the
execution of IRQportableStartKernel.
Upon return, miosix::kernel_started is set to be true and
IRQportableFinishKernelStartup is called immediately.

A motivation for this kind of flow could be that it allows running of
general purpose driver classes that would be ran either before or after
start of the kernel.
Probably these drivers may need to disable interrupts using
InterruptDisableLock in the case that they are initialized after
kernel's startup, while using InterruptDisableLock is error-prone when
the kernel_started flag is true and the kernel is not fully started yet
and the driver is started within IRQportableStartKernel.

Signed-off-by: Sasan Golchin <ahmad.golchin@mail.polimi.it>
Signed-off-by: Terraneo Federico <fede.tft@miosix.org>