Fix process context initialization on Cortex M3.
Inadvertently broken in 89aff2c1. ctxsave is just 9 elements, writing to the
10th element is a buffer overflow. Due to memory alignment effects, this buffer
overflow overwrote the following malloc chunk tag, creating a malloc chunk with
negative size. Havoc ensues as newlib malloc (which is just dlmalloc 2.6.5)
does not have any heap corruption checks. The negative chunk size is taken as
good, and when the corrupted chunk is freed or used, instead of updating the
next chunk's USED_PREV tag the malloc code corrupts the ctxsave structure.

Signed-off-by: Terraneo Federico <fede.tft@miosix.org>