Fix error code of isatty() when filesystem is disabled.
With filesystem disabled, only file descriptors 0, 1, 2 are open and accessible. Other file descriptors cannot be opened, so they should behave as invalid file descriptors. However for all other descriptors isatty() returned ENOTTY, which is not correct as this error code should only be returned when specifying a open and valid descriptor. The correct error code for a file descriptor which is not open is EBADF instead.

Signed-off-by: Terraneo Federico <fede.tft@miosix.org>