Fix to prevent some SD card brands from getting stuck in 4bit mode.
Some cards such as samsung evo plus microSD seem to have a bug where sometimes
the READY_FOR_DATA is never signaled. It seems to happen only when using 4 bit
mode and only in write commands. The issue seems unrelated to the board as it
happens both on an stm32f469ni_stm32f469i-disco and stm32f103ze_stm3210e-eval.
It is also unrelated to the recent Miosix development, as it is present at least
all the way back to Miosix v2.7.

When the bug happens, READY_FOR_DATA is stuck, and the card transitions from
Prg to Dis but never to Stby. As a workaround, considering the card as ready
when it enters the Dis state in waitForCardReady() passes the Miosix testsuite.

A way was also found to prevent the issue, but at a performance penalty: add
two delays in multipleBlockWrite a >=40us delay just before sending CMD25/CMD24,
and a >=30us delay after the DMA transfer is complete (before the code that
optionally sends CMD12). The reason why it works is still unknown, needs
further investigation.