Fixed `.map` file generation in CMake build system
Previously the map file name was hardcoded in the linker flags.
This was a problem when multiple programs (such as in the case
of processes) where build, the files would override one another
and only the last one would remain.

Now the map file gets the same path and name as the binary for
each target.