GCC build script: minor revamp of status reporting, logging, error handling.
All compilation output is redirected to the log files, to resolve the problem of not being able to track down in which build phase an error occurred. This also reduces noise on the terminal, speeding up the process a bit, at the cost of not being able to look at the files scrolling by. To compensate, more detailed status output is printed to stdout. Additionally, log file names have been changed to comprehensible ones rather than single letters. The code has also been slightly reformatted for legibility and better adherence to the column limit.

Signed-off-by: Terraneo Federico <fede.tft@miosix.org>